+++ src/client/ui/ui-workspace.mjs
@@ -1,13 +1,10 @@
 import UiWidget from './ui-widget.mjs'
-import {setScrollbarColor} from './utils.mjs'
 
 class UiWorkspace extends UiWidget {
 
     constructor(options) {
 
         super(options)
-
-        setScrollbarColor(this.container)
 
     }